+++ api/src/main/java/org/svip/api/controller/SVIPApiController.java
@@ -15,7 +15,10 @@ import org.svip.api.repository.SBOMFileRepository;
 import org.svip.api.utils.Converter;
 import org.svip.api.utils.Utils;
 import org.svip.sbom.builder.objects.SVIPSBOMBuilder;
+import org.svip.sbom.model.interfaces.generics.Component;
 import org.svip.sbom.model.interfaces.generics.SBOM;
+import org.svip.sbom.model.interfaces.generics.SBOMPackage;
+import org.svip.sbom.model.interfaces.schemas.SPDX23.SPDX23File;
 import org.svip.sbom.model.objects.CycloneDX14.CDX14SBOM;
 import org.svip.sbom.model.objects.SPDX23.SPDX23SBOM;
 import org.svip.sbom.model.objects.SVIPSBOM;
@@ -31,6 +34,13 @@ import org.svip.sbomgeneration.parsers.ParserController;
 import org.svip.sbomgeneration.serializers.SerializerFactory;
 import org.svip.sbomgeneration.serializers.deserializer.Deserializer;
 import org.svip.sbomgeneration.serializers.serializer.Serializer;
+import org.svip.sbomvex.VEXResult;
+import org.svip.sbomvex.database.NVDClient;
+import org.svip.sbomvex.database.OSVClient;
+import org.svip.sbomvex.database.interfaces.VulnerabilityDBClient;
+import org.svip.sbomvex.model.VEX;
+import org.svip.sbomvex.model.VEXType;
+import org.svip.sbomvex.vexstatement.VEXStatement;
 import org.svip.utils.VirtualPath;
 
 import java.util.*;
@@ -530,5 +540,102 @@ public class SVIPApiController {
 
         return Utils.encodeResponse(result.getContents());
 
+    }
+
+    /** USAGE Send GET request to /vex to generate a VEX Document for an SBOM
+     * The API will respond with an HTTP 200 a VEX object, and a hashmap of
+     * and errors that occurred
+     *
+     * @param id The id of the SBOM contents to retrieve.
+     * @param format the format of teh VEX Document
+     * @param client the api client to use (currently NVD or OSV)
+     * @return A new VEXResult of the VEX document and any errors that occurred
+     */
+    @GetMapping("/sboms/vex")
+    public ResponseEntity<VEXResult> vex(@RequestParam("id") long id,
+                                     @RequestParam("format") String format,
+                                     @RequestParam("client") String client){
+        SBOM sbom;
+        Deserializer d;
+
+        // Get the SBOM to be tested
+        Optional<SBOMFile> sbomFile = sbomFileRepository.findById(id);
+
+        // Check if it exists
+        if (sbomFile.isEmpty()) {
+            LOGGER.info("VEX /svip/sboms/vex?id=" + id + " - FILE NOT FOUND");
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        // Deserialize SBOM into JSON object
+        try{
+            d = SerializerFactory.createDeserializer(sbomFile.get().getContents());
+            sbom = d.readFromString(sbomFile.get().getContents());
+        } catch (JsonProcessingException e ){
+            LOGGER.info("Failed to deserialize SBOM content, may be an unsupported format");
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        } catch (Exception e){
+            LOGGER.info("Deserialization Error");
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
+        VulnerabilityDBClient vc;
+        // check that user entered a valid database client
+        switch(client.toLowerCase()){
+            case "osv" -> vc = new OSVClient();
+            case "nvd" -> vc = new NVDClient();
+            default -> {
+                LOGGER.info("VEX /svip/sboms/vex?client=" + client + " - INVALID CLIENT");
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+        }
+
+        // create new VEX builder and add sbom fields
+        VEX.Builder vb = new VEX.Builder();
+        String creationTime = String.valueOf(java.time.LocalDateTime.now());
+        vb.setVEXIdentifier(sbom.getName());
+        vb.setDocVersion("1.0");
+        vb.setTimeFirstIssued(creationTime);
+        vb.setTimeLastUpdated(creationTime);
+
+        // check that user entered a valid format
+        switch(format.toLowerCase()){
+            case "cyclonedx" -> {
+                vb.setOriginType(VEXType.CYCLONE_DX);
+                vb.setSpecVersion("1.4");
+            }
+            case "csaf" -> {
+                vb.setOriginType(VEXType.CSAF);
+                vb.setSpecVersion("2.0");
+            }
+            default -> {
+                LOGGER.info("VEX /svip/sboms/vex?format=" + format + " - INVALID VEX FORMAT");
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+        }
+
+        HashMap<String, String> error = new HashMap<>();
+        // add vex statements and/or errors for every component
+        for(Component c : sbom.getComponents()){
+            if(c instanceof SBOMPackage){
+                try{
+                    List<VEXStatement> statements = vc.getVEXStatements((SBOMPackage) c);
+                    if(!statements.isEmpty())
+                        for(VEXStatement vs : statements)
+                            vb.addVEXStatement(vs);
+                } catch (Exception e) {
+                    error.put(c.getName(), e.getMessage());
+                }
+            }
+        }
+
+        VEX vex = vb.build();
+
+        // Log
+        LOGGER.info("VEX /svip/sboms/vex?id=" + id + " - VEX CREATED: " + sbomFile.get().getFileName());
+
+        // Return VEXResult
+        return new ResponseEntity<>(
+                new VEXResult(vex, error), HttpStatus.OK);
     }
 }